Replaced SimpleInput with WizardInput Just a small refactoring, no new fucntionality.
