+++ Assets/Resources/ServerObjects/WizardPlayerDummy.prefab
@@ -343,7 +343,7 @@ Canvas:
   m_AdditionalShaderChannelsFlag: 0
   m_SortingLayerID: 0
   m_SortingOrder: 0
-  m_TargetDisplay: 1
+  m_TargetDisplay: 0
 --- !u!224 &224666393160540272
 RectTransform:
   m_ObjectHideFlags: 1
