--- Assets/Scripts/Player/PlayerControls/SimpleInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 02aee576905784417bf0acebcdc0ea8e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
